Move GoBackButton component to a shared ui directory and update import path in NotFound
